+++ lib/agent/index.js
@@ -1,12 +1,17 @@
 'use strict'
-const { InvalidArgumentError } = require('./core/errors')
-const Pool = require('./client-pool')
-const Client = require('./core/client')
+
+const { InvalidArgumentError } = require('../core/errors')
+const Pool = require('../client-pool')
+const Client = require('../core/client')
 const EventEmitter = require('events')
+const util = require('../core/util')
+const assert = require('assert')
+const RedirectHandler = require('./redirect')
 
 const kOnConnect = Symbol('onConnect')
 const kOnDisconnect = Symbol('onDisconnect')
 const kCache = Symbol('cache')
+const kMaxRedirections = Symbol('maxRedirections')
 const kFactory = Symbol('factory')
 
 function defaultFactory (origin, opts) {
@@ -16,13 +21,18 @@ function defaultFactory (origin, opts) {
 }
 
 class Agent extends EventEmitter {
-  constructor ({ factory = defaultFactory, ...opts } = {}) {
+  constructor ({ factory = defaultFactory, maxRedirections = 0, ...opts } = {}) {
     super()
 
     if (typeof factory !== 'function') {
       throw new InvalidArgumentError('factory must be a function.')
     }
 
+    if (!Number.isInteger(maxRedirections) || maxRedirections < 0) {
+      throw new InvalidArgumentError('maxRedirections must be a positive number')
+    }
+
+    this[kMaxRedirections] = maxRedirections
     this[kFactory] = (origin) => factory(origin, opts)
     this[kCache] = new Map()
 
@@ -65,7 +75,49 @@ class Agent extends EventEmitter {
       throw new InvalidArgumentError('Origin must be a non-empty string.')
     }
 
-    this.get(opts.origin).dispatch(opts, handler)
+    const { maxRedirections = this[kMaxRedirections] } = opts
+
+    if (!Number.isInteger(maxRedirections) || maxRedirections < 0) {
+      throw new InvalidArgumentError('maxRedirections must be a positive number')
+    }
+
+    const client = this.get(opts.origin)
+
+    if (!maxRedirections) {
+      client.dispatch(opts, handler)
+      return
+    }
+
+    if (util.isStream(opts.body) && util.bodyLength(opts.body) !== 0) {
+      // TODO (fix): Provide some way for the user to cache the file to e.g. /tmp
+      // so that it can be dispatched again?
+      // TODO (fix): Do we need 100-expect support to provide a way to do this properly?
+      client.dispatch(opts, handler)
+      return
+    }
+
+    opts = { ...opts }
+
+    const redirectHandler = new RedirectHandler({ agent: this, opts, handler })
+
+    if (util.isStream(opts.body)) {
+      opts.body
+        .on('data', /* istanbul ignore next */ function () {
+          /* istanbul ignore next */
+          assert(false)
+        })
+        .on('error', function (err) {
+          redirectHandler.destroy(err)
+        })
+        .on('end', function () {
+          util.destroy(this)
+        })
+      opts.body = null
+    }
+
+    opts.maxRedirections = maxRedirections - 1
+
+    client.dispatch(opts, redirectHandler)
   }
 
   close () {
