+++ lib/agent/redirect.js
@@ -1,11 +1,9 @@
-const { InvalidArgumentError } = require('./core/errors')
-const util = require('./core/util')
-const { getGlobalAgent } = require('./agent')
-const assert = require('assert')
+'use strict'
 
-const kMaxRedirections = Symbol('maxRedirections')
+const { InvalidArgumentError } = require('../core/errors')
+const util = require('../core/util')
+
 const kAgent = Symbol('agent')
-const kRedirectTag = Symbol('redirectTag')
 
 class RedirectHandler {
   constructor ({ agent, opts, handler }) {
@@ -94,71 +92,6 @@ class RedirectHandler {
   }
 }
 
-class RedirectAgent {
-  constructor ({ maxRedirections = 10, agent } = {}) {
-    if (!Number.isInteger(maxRedirections) || maxRedirections < 0) {
-      throw new InvalidArgumentError('maxRedirections must be a positive number')
-    }
-
-    if (agent != null && typeof agent.dispatch !== 'function') {
-      throw new InvalidArgumentError('invalid opts.agent')
-    }
-
-    this[kMaxRedirections] = maxRedirections
-    this[kAgent] = agent
-    this[kRedirectTag] = Symbol('redirectTag')
-  }
-
-  dispatch (opts, handler) {
-    const { maxRedirections = this[kMaxRedirections] } = opts
-
-    if (!Number.isInteger(maxRedirections) || maxRedirections < 0) {
-      throw new InvalidArgumentError('maxRedirections must be a positive number')
-    }
-
-    opts = { ...opts }
-
-    assert(!opts[kRedirectTag] || (opts[kRedirectTag] === this[kRedirectTag]), 'circular dispatch')
-    opts[kRedirectTag] = this[kRedirectTag]
-
-    const { [kAgent]: agent = getGlobalAgent() } = this
-
-    if (!maxRedirections) {
-      agent.dispatch(opts, handler)
-      return
-    }
-
-    if (util.isStream(opts.body) && util.bodyLength(opts.body) !== 0) {
-      // TODO (fix): Provide some way for the user to cache the file to e.g. /tmp
-      // so that it can be dispatched again?
-      // TODO (fix): Do we need 100-expect support to provide a way to do this properly?
-      agent.dispatch(opts, handler)
-      return
-    }
-
-    const redirectHandler = new RedirectHandler({ agent: this, opts, handler })
-
-    if (util.isStream(opts.body)) {
-      opts.body
-        .on('data', /* istanbul ignore next */ function () {
-          /* istanbul ignore next */
-          assert(false)
-        })
-        .on('error', function (err) {
-          redirectHandler.destroy(err)
-        })
-        .on('end', function () {
-          util.destroy(this)
-        })
-      opts.body = null
-    }
-
-    opts.maxRedirections = maxRedirections - 1
-
-    agent.dispatch(opts, redirectHandler)
-  }
-}
-
 function parseLocation (headers) {
   for (let i = 0; i < headers.length; i += 2) {
     if (headers[i].length === 8 && headers[i].toLowerCase() === 'location') {
@@ -196,4 +129,4 @@ function cleanRequestHeaders (headers, removeContent) {
   return ret
 }
 
-module.exports.RedirectAgent = RedirectAgent
+module.exports = RedirectHandler
